taiga-server/taiga/main/serializers.py
from rest_framework import serializers
from .models import UploadImageTest, UploadFile
from django.contrib.auth.password_validation import validate_password


class ImageSerializer(serializers.ModelSerializer):

    user = serializers.HiddenField(default=serializers.CurrentUserDefault())

    class Meta:
        model = UploadImageTest
        fields = '__all__'


class FileSerializer(serializers.ModelSerializer):

    user = serializers.HiddenField(default=serializers.CurrentUserDefault())

    class Meta:
        model = UploadFile

        fields = '__all__'


class ChangePasswordSerializer(serializers.Serializer):
    old_password = serializers.CharField(required=True)
    new_password = serializers.CharField(required=True)

    def validate_new_password(self, value):
        validate_password(value)
        return value